Update doors.lua
They where opening but with punch. I moved it to right click

Door placement and breaking was quite buggy.
I could place doors on top of eachother and half doors appeared
When I placed 2 doors next to eachother and I broke them 1 half a door kept standing

Placement was limitless I could place a door in every gap :)

Fixed those things.

When holding an orb and rightclick on the door it opens.(maybe you want some metadata at the door with username attached? so it opens only for the user? )

Placement are now only allowed when the ground has 2 free nodes next to eachother so it does not place in the ground anymore.

Breaking is fixed